requirements.txt
winrt-Windows.Foundation
winrt-Windows.Foundation.Collections
winrt-Windows.Media
winrt-Windows.Media.Control